strategy/requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
akshare==1.16.44
alpaca-py==0.39.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dashscope==1.22.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
ipykernel==6.29.5
ipython==8.32.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lib-pybroker==1.2.7
llvmlite==0.44.0
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mini-racer==0.12.4
missingno==0.5.2
mistune==3.1.2
msgpack==1.1.0
multidict==6.1.0
multimethod==2.0
multitasking==0.0.11
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodejs==0.1.1
notebook==6.5.4
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openai==1.65.3
openpyxl==3.1.5
optional-django==0.1.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.9
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
progressbar2==4.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-utils==3.9.1
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-futures==1.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sseclient-py==1.8.0
stack-data==0.6.3
ta-lib @ file:///D:/Projects/DevilProjects/FissionMachine/strategy/ta_lib-0.6.3-cp312-cp312-win_amd64.whl#sha256=fbf480e9e936ed983776c616f461f4d4511bb1d010424558db3f559aec10dcd4
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
visions==0.7.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
xlrd==2.0.1
yahooquery==2.3.7
yarl==1.18.3
yfinance==0.2.54
